backend/requirements.txt
# RoboLabel backend — pin versions for reproducible builds
Django>=5.0,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
Pillow>=10.0,<12
PyYAML>=6.0,<7
opencv-python-headless>=4.9,<5
numpy>=1.26,<3
